requirements.txt
python-dotenv>=1.0.0
google-genai>=1.0.0
google-api-python-client>=2.140.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
msal>=1.30.0
requests>=2.32.0
textual>=0.83.0
rich>=13.9.0
